renaming wx/qt4 registry_helper as _registry_helper
