+++ traitsui/testing/tester/qt4/_registry_helper.py
