+++ traitsui/testing/tester/qt4/implementation/editor_factory.py
@@ -9,7 +9,7 @@
 #  Thanks for using Enthought open source!
 #
 from traitsui.testing.tester import query
-from traitsui.testing.tester.qt4.registry_helper import (
+from traitsui.testing.tester.qt4._registry_helper import (
     register_editable_textbox_handlers,
 )
 from traitsui.qt4.editor_factory import ReadonlyEditor, TextEditor
