+++ traitsui/testing/tester/wx/_registry_helper.py
