+++ traitsui/testing/tester/wx/implementation/editor_factory.py
@@ -11,7 +11,7 @@
 from traitsui.wx.editor_factory import ReadonlyEditor, TextEditor
 from traitsui.testing.tester import query
 from traitsui.testing.tester.wx import _interaction_helpers
-from traitsui.testing.tester.wx.registry_helper import (
+from traitsui.testing.tester.wx._registry_helper import (
     register_editable_textbox_handlers,
 )
