+++ traitsui/testing/tester/wx/implementation/range_editor.py
@@ -16,7 +16,7 @@ from traitsui.wx.range_editor import (
 )
 
 from traitsui.testing.tester import locator
-from traitsui.testing.tester.wx import registry_helper
+from traitsui.testing.tester.wx import _registry_helper
 from traitsui.testing.tester.wx.common_ui_targets import (
     LocatedSlider,
     LocatedTextbox
@@ -50,7 +50,7 @@ def register(registry):
             solver=lambda wrapper, _: LocatedSlider(
                 slider=wrapper._target.control.slider),
         )
-    registry_helper.register_editable_textbox_handlers(
+    _registry_helper.register_editable_textbox_handlers(
         registry=registry,
         target_class=RangeTextEditor,
         widget_getter=lambda wrapper: wrapper._target.control,
