+++ traitsui/tests/editors/test_range_editor.py
@@ -29,10 +29,10 @@ def _register_simple_spin(registry):
     registry : TargetRegistry
         The registry being registered to.
     """
-    from traitsui.testing.tester.qt4 import registry_helper
+    from traitsui.testing.tester.qt4 import _registry_helper
     from traitsui.qt4.range_editor import SimpleSpinEditor
 
-    registry_helper.register_editable_textbox_handlers(
+    _registry_helper.register_editable_textbox_handlers(
         registry=registry,
         target_class=SimpleSpinEditor,
         widget_getter=lambda wrapper: wrapper._target.control.lineEdit(),
